src/backend/requirements.txt
Flask
Flask-Cors
gunicorn
mido
numpy
Pillow